# Few optimizations for FITS tables