I have FITS files with a lot of extensions, lot of tables with many columns (a few thousands sometimes ...) and it takes a significant time to load everything, so I tried to find some optimizations (btw snakeviz is very nice to look at profiling results !). 
The changes here could be seen a micro-optimizations, but they concern methods which are called many times, so it makes a difference in the end.

* Reference: Astropy 1.2.dev15480, Time to load my file : 2.8s
* Change on `_is_int` (ncalls=1621317). Time after : 2.64 s
* Change on `Coldefs.__getattr__` (ncalls=2673). Time after : 1.73s
* Change on `ColumnAttribute.__get__` (ncalls=1626425). Time after : 1.6s

Maybe @embray if you have some time ?