@saimn 2016-06-02
@pllim @MSeifert04 - With https://github.com/astropy/astropy/pull/4985/commits/3af359c50fba917fa721bdf78087a587cb5b37a1 I now cache the lists of attributes in `ColDefs` instead of calling too many times `__getattr__` (for `name`, `format`, ...). This results in a 2.5x speedup (just for this commit) ! I hope I handle correctly all the cases when the cache must be updated, at least the tests pass.

@embray - Maybe, if you want have a look ?

Master:

    ~/dev/astropy master*
    ❯ python -m timeit -n 5 -r 5 -s "from astropy.table import Table" "Table.read('table.fits')"
    5 loops, best of 5: 2.44 sec per loop

This branch:

    ~/dev/astropy fits-table-optims*
    ❯ python -m timeit -n 5 -r 5 -s "from astropy.table import Table" "Table.read('table.fits')"
    5 loops, best of 5: 485 msec per loop