@pllim 2016-06-02
Overall, the speed-up looks nice. Like you said, the tests passed. So unless someone takes this for a spin on a variety of FITS tables, it is good enough?