@MSeifert04 2016-09-06
@saimn Since this is quite a small PR could you squash the commits into one and include a changelog entry (just to mention the improved runtime)? Maybe wait until the tests have finished to see if anything went wrong.

I'm not sure this is strictly necessary here so maybe wait if someone else wants to second that.

I'm really interested about the fitsrec-changes. I've experimented with them before and always got several failures, so I'm really interested to see if it works out. :+1: