@saimn 2016-09-06
@MSeifert04 - I can probably squash 2 or 3 commits for things that are modified in another commit, but there are a number of commits I would like to keep as they address specific things and contain some information on the change.