@saimn 2016-09-06
Ok, squashed some commits, and added changelog.