@saimn 2016-09-07
The build failure seems unrelated (something with a samp hub process: https://travis-ci.org/astropy/astropy/jobs/157994577#L2451).