@embray 2016-09-07
Great--thanks for putting up with my nitpicks.  I think this is good!