@taldcroft 2016-10-14
@saimn - I'm not sure if it was exactly this PR, but reading short wide tables is *way* faster for me now, as in a factor of 8!  (Comparing 1.2.1 to current master 1.3.dev16204).  This is awesome and therefore closes #4609.