@taldcroft 2016-10-14
Specifically, for this table I get the factor of 8 improvement:
```
import numpy as np
from astropy.table import Table

ncols = 800
nrows = 10
cols = [np.random.uniform(size=nrows) for _ in xrange(ncols)]
dat = Table(cols)

dat.write('test_speed.fits', overwrite=True)
```