@saimn 2016-10-14
@taldcroft - This is typically the use case I was addressing here, tables with a many columns, so I'm glad it is useful to someone else :)