@embray 2016-10-17
Thanks @taldcroft for the testimonial--I bet it was this PR!